refactor: 최단경로 조회 인터페이스 도메인 패키지로 이동
